pyproject.toml
[project]
name = "legobot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "feetech-servo-sdk>=1.0.0",
    "pyserial>=3.5",
    "pyyaml>=6.0.3",
    "ruff>=0.14.10",
]
